Restrict function syntax
The new function syntax mandates that a return type is specified, but
leaves optional the object type (whether it is a function, method,
override etc).

The shorthand function-syntax is just a constrained version of the full
declaration syntax, and is only intended to improve readability (by
enforcing uniformity) as well as avoiding ambiguity in the parser.

If more flexibility is needed for defining functions (unlikely) a user
can always revert to the full declaration syntax, which means defining
the function parameters as a collection.